스프링 06일차 : 댓글 수정 구현

+++ spring projects/spring_t/src/main/java/kr/kh/spring/service/CommentService.java
@@ -51,6 +51,27 @@ public class CommentService {
 		if(user == null) {
 			return false;
 		}
+		if(!isWriter(cm_num, user.getMe_id())) {
+			return false;
+		}
 		return commentDao.deleteComment(cm_num);
+	}
+
+	public boolean updateComment(CommentVO comment, MemberVO user) {
+		if(user == null || comment == null) {
+			return false;
+		}
+		if(!isWriter(comment.getCm_num(), user.getMe_id())) {
+			return false;
+		}
+		return commentDao.updateComment(comment);
+	}
+	public boolean isWriter(int cm_num, String me_id) {
+		CommentVO comment = commentDao.selectComment(cm_num);
+		if(comment == null ) {
+			return false;
+		}
+		return comment.getCm_me_id().equals(me_id);
+
 	}
 }
